# src/config/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Perfume Shop API",
    "description": "RESTful API for a perfume and cologne e-commerce platform with multi-variant product support, allowing flexible volume-based pricing and weight-based sales",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.yourdomain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/{orderNumber}/status": {
      "patch": {
        "tags": [
          "Admin - Orders"
        ],
        "summary": "Change order status",
        "description": "Update order status and add status history note",
        "parameters": [
          {
            "name": "orderNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order number"
          }
        ],
        "responses": {
          "200": {
            "description": "Order status updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orderNumber": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required or not admin"
          },
          "404": {
            "description": "Order not found or has been cancelled"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "approved",
                      "rejected",
                      "shipped",
                      "delivered",
                      "cancelled"
                    ],
                    "example": "shipped",
                    "description": "New order status"
                  },
                  "note": {
                    "type": "string",
                    "example": "Order has been packed",
                    "description": "Optional note to add to status history (max 200 characters)"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        }
      }
    },
    "/{orderNumber}/cancel": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Cancel an order",
        "description": "Cancel a pending order and restore product stock",
        "parameters": [
          {
            "name": "orderNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order number"
          }
        ],
        "responses": {
          "200": {
            "description": "Order cancelled successfully"
          },
          "400": {
            "description": "Order already cancelled or not in pending status"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Order not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/{slug}/comments": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product comments",
        "description": "Retrieve all approved comments for a specific product",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product slug"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comments retrieved successfully"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/{slug}/comment": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a product comment",
        "description": "Submit a comment and rating for a product",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product slug"
          }
        ],
        "responses": {
          "201": {
            "description": "Comment submitted successfully"
          },
          "400": {
            "description": "Validation error or already commented"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "example": "Great product! Highly recommended."
                  },
                  "rating": {
                    "type": "integer",
                    "example": 5,
                    "minimum": 1,
                    "maximum": 5
                  }
                },
                "required": [
                  "content",
                  "rating"
                ]
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    }
  }
}